Add support for ICN file format
As described here:
https://wiki.xxiivv.com/site/icn_format.html

This is exactly the same as the 1BPP NES/Monochrome format described in
the link on this issue as well: https://github.com/zigimg/zigimg/issues/30

There is some weirdness with a tiled, linear bitmap format in that image dimensions aren't encoded in the file - they have to be known beforehand. We could make our best guess based on the file size and assume it's square, but some graphics software that can read/write ICN files can handle non-square sizes.